Remove json-schema gem and mock replacement usage.
