+++ app/models/callbacks/api/v1/schema_callbacks.rb
@@ -1,0 +1,29 @@
+module API
+  module V1
+    module SchemaCallbacks
+      extend ActiveSupport::Concern
+
+      included do
+        after_commit :save_version
+        after_commit :touch_versions, on: :update
+      end
+
+      private
+
+      def save_version
+        with_lock do
+          versions.create(
+            schema_ref: schema_ref,
+            schema: schema,
+            example: example,
+            version_number: versions.count + 1
+          )
+        end
+      end
+
+      def touch_versions
+        versions.each(&:touch)
+      end
+    end
+  end
+end
