+++ config/initializers/json_schema.rb
@@ -1,0 +1,2 @@
+# Load patch lib.
+require 'dry-json-schemas/json-schema'
